requirements.txt
# PreenCut Requirements
# Web Framework and API
fastapi==0.115.13
uvicorn[standard]==0.32.1
gradio==5.34.2

# AI and Machine Learning
torch==2.7.1
torchaudio==2.7.1
whisperx==3.3.4
faster-whisper==1.1.1
openai==1.91.0

# Media Processing
ffmpeg-python==0.2.0

# Data Handling and Validation
pydantic==2.10.4
pydantic-settings==2.7.0

# Environment and Configuration
python-dotenv==1.0.1

# Development and Testing (optional)
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# Logging and Monitoring
structlog==24.4.0

# Utilities
pathlib2==2.3.7
typing-extensions==4.12.2
